MOVE-4266-Added support for custom claims and jwk (#239)
Added support for custom claims, needed for fetching systemuser for DPV altinn 3. Added support for fetching token using jwk instead of certificate, used for testing DPO and DPV with tenor testorg.
[MOVE-4266-SupportForJwkAndCustomClaims 527ab10] Added support for custom claims, needed for fetching systemuser for DPV altinn 3. Added support for fetching token using jwk instead of certificate, used for testing DPO and DPV with tenor testorg.

Removed spring-boot-starter-web, dosnt seem to be in use and it creates problems for DPV proxy client when the proxy uses common security